pyproject.toml
[project]
name = "etl-crypto"
version = "1.0.0"
description = "Real-Time ETL pipeline for streaming Crypto Data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "asyncio>=4.0.0",
    "binance-sdk-spot>=5.0.0",
    "kafka-python>=2.3.0",
    "pandas>=2.3.3",
    "pyspark>=4.1.0",
    "websocket-client>=1.9.0",
    "websockets>=15.0.1",
]
